Rename Matcher to SmallPatternMatcher
This is the first step towards introducing large patterns. As we will
have two different kinds of patterns we need to have more specific names!
