+++ src/patterns/small/mod.rs
