+++ src/patterns/small/pattern/mod.rs
