+++ src/patterns/small/pattern/test.rs
