+++ src/patterns/small/point/test.rs
